add teams, systems, and users pages
