+++ components/GetUsers.tsx
@@ -5,7 +5,7 @@ import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
-import Title from './Title';
+import Title from './common/Title';
 import Cookies from 'js-cookie';
 import axios from 'axios';
